requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
langchain-community
langchain-groq
langchain-text-splitters
streamlit
requests
python-dotenv
pypdf
faiss-cpu
fastembed
python-multipart